refactor: rename file to match contained class
