+++ backend/src/main/game/monster/abstract_monster_card.py
