+++ backend/src/main/game/monster/concrete_monster_cards/rotting.py
@@ -1,11 +1,10 @@
-from backend.src.main.game.monster.random_monster_card import AbstractMonsterCard
+from backend.src.main.game.monster.abstract_monster_card import AbstractMonsterCard
 from backend.src.main.game.monster.values import DungeonCardValues, NumberedRoomTileValues
 
 
 class Rotting(AbstractMonsterCard):  # pylint: disable=too-few-public-methods
 
     def __init__(self):
-
         map_values = {NumberedRoomTileValues.ONE: DungeonCardValues.MONSTER,
                       NumberedRoomTileValues.THREE: DungeonCardValues.MONSTER,
                       NumberedRoomTileValues.FOUR: DungeonCardValues.MONSTER,
