+++ backend/src/main/game/room/constructed_room.py
@@ -1,4 +1,4 @@
-from backend.src.main.game.monster.random_monster_card import AbstractMonsterCard
+from backend.src.main.game.monster.abstract_monster_card import AbstractMonsterCard
 from backend.src.main.game.monster.values import NumberedRoomTileValues
 from backend.src.main.game.room.room import AbstractRoomCard
 from backend.src.main.game.tile.tile import Tile
