+++ backend/src/main/game/room/waypoint/waypoint_a.py
