+++ backend/src/main/game/room/waypoint/waypoint_b.py
