+++ backend/src/test/game/monster/test_cutthroat.py
@@ -1,7 +1,6 @@
 import pytest
-
+from backend.src.main.game.monster.abstract_monster_card import AbstractMonsterCard
 from backend.src.main.game.monster.concrete_monster_cards.cutthroat import Cutthroat
-from backend.src.main.game.monster.random_monster_card import AbstractMonsterCard
 from backend.src.main.game.monster.values import DungeonCardValues, NumberedRoomTileValues
