+++ backend/src/test/game/tile/test_tile_geometry.py
@@ -6,7 +6,7 @@ import pytest
 from backend.src.main.game.monster.values import UniqueDungeonCardValues
 from backend.src.main.game.room.concrete_room_cards.burrow import Burrow
 from backend.src.main.game.room.concrete_room_cards.tunnel import Tunnel
-from backend.src.main.game.room.waypoint.waypoint_b_pojo import WaypointB
+from backend.src.main.game.room.waypoint.waypoint_b import WaypointB
 from backend.src.main.game.tile.tile import Tile
 from backend.src.main.game.tile.tile_geometry import TileGeometry
